Aggiunta ruoli

+++ controller/lecture_controller.go
@@ -1,6 +1,7 @@
 package controller
 
 import (
+	en "go-server/EnumRole"
 	"go-server/data/request"
 	"go-server/data/response"
 	"go-server/helper"
@@ -22,27 +23,35 @@ func NewLecturesController(lectureService lecture.LectureService) *LectureContro
 }
 
 func (controller *LectureController) Create(writer http.ResponseWriter, requests *http.Request, params httprouter.Params) {
-	valid := service.CheckToken(requests)
-	if valid {
-		courseRequest := request.LectureCreateRequest{}
-		helper.ReadRequestBody(requests, &courseRequest)
-		result, errx := controller.lectureService.Create(requests.Context(), courseRequest)
-		if result {
-			WebResponse := helper.WebResponse{
-				Status: "ok",
-				Data:   courseRequest,
+	if !en.IsStudent(requests) {
+		valid := service.CheckToken(requests)
+		if valid {
+			courseRequest := request.LectureCreateRequest{}
+			helper.ReadRequestBody(requests, &courseRequest)
+			result, errx := controller.lectureService.Create(requests.Context(), courseRequest)
+			if result {
+				WebResponse := helper.WebResponse{
+					Status: "ok",
+					Data:   courseRequest,
+				}
+				helper.WriteResponse(writer, WebResponse, http.StatusCreated)
+			} else {
+				webRepo := helper.WebResponse{
+					//Status: "Error during create",
+					Status: errx.Error(),
+				}
+				helper.WriteResponse(writer, webRepo, http.StatusConflict)
 			}
-			helper.WriteResponse(writer, WebResponse, http.StatusCreated)
 		} else {
 			webRepo := helper.WebResponse{
-				//Status: "Error during create",
-				Status: errx.Error(),
+				Status: "Error token not valid",
 			}
-			helper.WriteResponse(writer, webRepo, http.StatusConflict)
+
+			helper.WriteResponse(writer, webRepo, http.StatusUnauthorized)
 		}
 	} else {
 		webRepo := helper.WebResponse{
-			Status: "Error token not valid",
+			Status: "This is not your role",
 		}
 
 		helper.WriteResponse(writer, webRepo, http.StatusUnauthorized)
@@ -96,39 +105,48 @@ func (controller *LectureController) FindByIds(writer http.ResponseWriter, reque
 }
 
 func (controller *LectureController) Update(writer http.ResponseWriter, requests *http.Request, params httprouter.Params) {
-	valid := service.CheckToken(requests)
-	if valid {
-		lectureUpdate := request.LectureUpdateRequest{}
-		helper.ReadRequestBody(requests, &lectureUpdate)
-		name := params.ByName("name")
-		if name == "" {
-			webRepo := helper.WebResponse{
-				Status: "Error",
-				Data:   "Name is required",
+	if !en.IsStudent(requests) {
+		valid := service.CheckToken(requests)
+		if valid {
+			lectureUpdate := request.LectureUpdateRequest{}
+			helper.ReadRequestBody(requests, &lectureUpdate)
+			name := params.ByName("name")
+			if name == "" {
+				webRepo := helper.WebResponse{
+					Status: "Error",
+					Data:   "Name is required",
+				}
+				helper.WriteResponse(writer, webRepo, http.StatusBadRequest)
+				return
 			}
-			helper.WriteResponse(writer, webRepo, http.StatusBadRequest)
-			return
-		}
-		result, errx := controller.lectureService.Update(requests.Context(), lectureUpdate, name)
-		if result {
-			webRepo := helper.WebResponse{
+			result, errx := controller.lectureService.Update(requests.Context(), lectureUpdate, name)
+			if result {
+				webRepo := helper.WebResponse{
 
-				Status: "ok",
-				Data:   nil,
+					Status: "ok",
+					Data:   nil,
+				}
+				helper.WriteResponse(writer, webRepo, http.StatusNoContent)
+			} else {
+				webRepo := helper.WebResponse{
+
+					//Status: "Error during update",
+					Status: errx.Error(),
+				}
+				helper.WriteResponse(writer, webRepo, http.StatusNotFound)
 			}
-			helper.WriteResponse(writer, webRepo, http.StatusNoContent)
 		} else {
 			webRepo := helper.WebResponse{
 
-				//Status: "Error during update",
-				Status: errx.Error(),
+				Status: "Error token not valid",
 			}
-			helper.WriteResponse(writer, webRepo, http.StatusNotFound)
+
+			helper.WriteResponse(writer, webRepo, http.StatusUnauthorized)
 		}
 	} else {
 		webRepo := helper.WebResponse{
 
-			Status: "Error token not valid",
+			Status: "This is not your role",
 		}
 
 		helper.WriteResponse(writer, webRepo, http.StatusUnauthorized)
@@ -136,34 +154,42 @@ func (controller *LectureController) Update(writer http.ResponseWriter, requests
 }
 
 func (controller *LectureController) Delete(writer http.ResponseWriter, requests *http.Request, params httprouter.Params) {
-	valid := service.CheckToken(requests)
-	if valid {
-		name := params.ByName("name")
-		if name == "" {
-			webRepo := helper.WebResponse{
-				Status: "Error",
-				Data:   "Name is required",
+	if !en.IsStudent(requests) {
+		valid := service.CheckToken(requests)
+		if valid {
+			name := params.ByName("name")
+			if name == "" {
+				webRepo := helper.WebResponse{
+					Status: "Error",
+					Data:   "Name is required",
+				}
+				helper.WriteResponse(writer, webRepo, http.StatusBadRequest)
+				return
 			}
-			helper.WriteResponse(writer, webRepo, http.StatusBadRequest)
-			return
-		}
-		result, errx := controller.lectureService.Delete(requests.Context(), name)
-		if result {
-			webRepo := helper.WebResponse{
-				Status: "ok",
-				Data:   nil,
+			result, errx := controller.lectureService.Delete(requests.Context(), name)
+			if result {
+				webRepo := helper.WebResponse{
+					Status: "ok",
+					Data:   nil,
+				}
+				helper.WriteResponse(writer, webRepo, http.StatusNoContent)
+			} else {
+				webRepo := helper.WebResponse{
+					//Status: "Error during delete",
+					Status: errx.Error(),
+				}
+				helper.WriteResponse(writer, webRepo, http.StatusNotFound)
 			}
-			helper.WriteResponse(writer, webRepo, http.StatusNoContent)
 		} else {
 			webRepo := helper.WebResponse{
-				//Status: "Error during delete",
-				Status: errx.Error(),
+				Status: "Error token not valid",
 			}
-			helper.WriteResponse(writer, webRepo, http.StatusNotFound)
+
+			helper.WriteResponse(writer, webRepo, http.StatusUnauthorized)
 		}
 	} else {
 		webRepo := helper.WebResponse{
-			Status: "Error token not valid",
+			Status: "This is not your role",
 		}
 
 		helper.WriteResponse(writer, webRepo, http.StatusUnauthorized)
